use path helper function
