+++ src/components/fields/schemaFields/v1/SchemaField.stories.tsx
@@ -21,7 +21,7 @@ import { Form, Formik } from "formik";
 import SchemaField from "./SchemaField";
 import { SchemaFieldProps } from "@/components/fields/schemaFields/propTypes";
 import { Button } from "react-bootstrap";
-import { lastIndexOf } from "lodash";
+import { getFieldNamesFromPathString } from "@/runtime/pathHelpers";
 
 export default {
   title: "Fields/v1/SchemaField",
@@ -31,7 +31,7 @@ export default {
 const Template: ComponentStory<
   React.FunctionComponent<SchemaFieldProps & { defaultValue: unknown }>
 > = (args) => {
-  const fieldName = args.name.slice(lastIndexOf(args.name, ".") + 1);
+  const fieldName = getFieldNamesFromPathString(args.name)[1];
   return (
     <Formik
       initialValues={{
